+++ YNDropDownMenu/Classes/YNDropDownView.swift
@@ -18,7 +18,12 @@ open class YNDropDownView: UIView {
         self.delegate?.hideMenu()
     }
     
-    open func changeMenuTitleAt(index: Int, title: String) {
-        self.delegate?.changeMenuTitleAt(index: index, title: title)
+    
+    open func changeMenu(title: String, at index: Int) {
+        self.delegate?.changeMenu(title: title, at: index)
+    }
+    
+    open func changeView(view: UIView, at index: Int) {
+        self.delegate?.changeView(view: view, at: index)
     }
 }